Create CommonPort and EtherPort classes representing Ethernet specific  port functionality (including automotive) and functionality common  across all media These classes are pulled from the IEEE1588Port class which has been removed
This will reduce the amount of code duplication when support for 802.11
 media is added
