+++ daemons/gptp/common/ap_message.cpp
@@ -26,7 +26,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ieee1588.hpp>
 #include <avbts_clock.hpp>
 #include <avbap_message.hpp>
-#include <avbts_port.hpp>
+#include <ether_port.hpp>
 #include <avbts_ostimer.hpp>
 #include <gptp_log.hpp>
 
@@ -55,11 +55,11 @@ APMessageTestStatus::~APMessageTestStatus()
 {
 }
 
-APMessageTestStatus::APMessageTestStatus(IEEE1588Port * port)
+APMessageTestStatus::APMessageTestStatus( EtherPort *port )
 {
 }
 
-void APMessageTestStatus::sendPort(IEEE1588Port * port)
+void APMessageTestStatus::sendPort( EtherPort * port )
 {
 	static uint16_t sequenceId = 0;
