+++ daemons/gptp/common/avbts_message.hpp
@@ -235,9 +235,10 @@ protected:
  public:
 	/**
 	 * @brief Creates the PTPMessageCommon interface
-	 * @param port IEEE1588Port where the message interface is connected to.
+	 * @param port EtherPort where the message interface is
+	 * connected to.
 	 */
-	PTPMessageCommon(IEEE1588Port * port);
+	PTPMessageCommon( CommonPort *port );
 	/**
 	 * @brief Destroys PTPMessageCommon interface
 	 */
@@ -359,7 +360,7 @@ protected:
 	 * @param  port IEEE1588 port
 	 * @return void
 	 */
-	virtual void processMessage(IEEE1588Port * port);
+	virtual void processMessage( EtherPort *port );
 
 	/**
 	 * @brief  Builds PTP common header
@@ -369,7 +370,8 @@ protected:
 	void buildCommonHeader(uint8_t * buf);
 
 	friend PTPMessageCommon *buildPTPMessage
-	(char *buf, int size, LinkLayerAddress * remote, IEEE1588Port * port);
+	( char *buf, int size, LinkLayerAddress * remote,
+	  EtherPort *port );
 };
 
 /*Exact fit. No padding*/
@@ -505,7 +507,7 @@ class PTPMessageAnnounce:public PTPMessageCommon {
 	 /**
 	  * @brief Creates the PTPMessageAnnounce interface
 	  */
-	 PTPMessageAnnounce(IEEE1588Port * port);
+	 PTPMessageAnnounce( CommonPort * port );
 
 	 /**
 	  * @brief Destroys the PTPMessageAnnounce interface
@@ -573,22 +575,25 @@ class PTPMessageAnnounce:public PTPMessageCommon {
 
 	/**
 	 * @brief  Processes PTP message
-	 * @param  port IEEE1588Port
+	 * @param  port EtherPort
 	 * @return void
 	 */
-	void processMessage(IEEE1588Port * port);
+	void processMessage( EtherPort *port );
 
 	/**
-	 * @brief  Assembles PTPMessageAnnounce message on the IEEE1588Port payload
-	 * @param  port IEEE1588Port where the message will be assembled
+	 * @brief  Assembles PTPMessageAnnounce message on the
+	 * EtherPort payload
+	 * @param  port EtherPort where the message will be
+	 * assembled
 	 * @param  destIdentity [in] Destination PortIdentity
 	 * @return void
 	 */
-	void sendPort(IEEE1588Port * port, PortIdentity * destIdentity);
+	void sendPort
+	( CommonPort *port, PortIdentity *destIdentity);
 
-	friend PTPMessageCommon *buildPTPMessage(char *buf, int size,
-						 LinkLayerAddress * remote,
-						 IEEE1588Port * port);
+	friend PTPMessageCommon *buildPTPMessage
+	( char *buf, int size, LinkLayerAddress *remote,
+	  EtherPort *port );
 };
 
 /**
@@ -602,9 +607,9 @@ class PTPMessageSync : public PTPMessageCommon {
  public:
 	/**
 	 * @brief Default constructor. Creates PTPMessageSync
-	 * @param port IEEE1588Port
+	 * @param port EtherPort
 	 */
-	PTPMessageSync(IEEE1588Port * port);
+	PTPMessageSync( EtherPort *port );
 
 	/**
 	 * @brief Destroys PTPMessageSync interface
@@ -613,10 +618,10 @@ class PTPMessageSync : public PTPMessageCommon {
 
 	/**
 	 * @brief  Processes PTP messages
-	 * @param  port [in] IEEE1588Port
+	 * @param  port [in] EtherPort
 	 * @return void
 	 */
-	void processMessage(IEEE1588Port * port);
+	void processMessage( EtherPort *port );
 
 	/**
 	 * @brief  Gets origin timestamp value
@@ -627,15 +632,19 @@ class PTPMessageSync : public PTPMessageCommon {
 	}
 
 	/**
-	 * @brief  Assembles PTPMessageSync message on the IEEE1588Port payload
-	 * @param  port IEEE1588Port where the message will be assembled
+	 * @brief  Assembles PTPMessageSync message on the
+	 * EtherPort payload
+	 * @param  port EtherPort where the message will be
+	 * assembled
 	 * @param  destIdentity [in] Destination PortIdentity
 	 * @return void
 	 */
-	void sendPort(IEEE1588Port * port, PortIdentity * destIdentity);
+	void sendPort
+	(EtherPort *port, PortIdentity *destIdentity );
 
 	friend PTPMessageCommon *buildPTPMessage
-	(char *buf, int size, LinkLayerAddress * remote, IEEE1588Port * port);
+	( char *buf, int size, LinkLayerAddress * remote,
+	  EtherPort *port );
 };
 
 /* Exact fit. No padding*/
@@ -841,22 +850,26 @@ public:
 	/**
 	 * @brief Builds the PTPMessageFollowUP object
 	 */
-	PTPMessageFollowUp(IEEE1588Port * port);
+	PTPMessageFollowUp( EtherPort *port );
 
 	/**
-	 * @brief  Assembles PTPMessageFollowUp message on the IEEE1588Port payload
-	 * @param  port IEEE1588Port where the message will be assembled
+	 * @brief  Assembles PTPMessageFollowUp message on the
+	 * EtherPort payload
+	 * @param  port EtherPort where the message will be
+	 * assembled
 	 * @param  destIdentity [in] Destination PortIdentity
 	 * @return void
 	 */
-	void sendPort(IEEE1588Port * port, PortIdentity * destIdentity);
+	void sendPort
+	( EtherPort *port, PortIdentity *destIdentity );
 
 	/**
 	 * @brief  Processes PTP messages
-	 * @param  port [in] IEEE1588Port
+	 * @param  port [in] EtherPort
 	 * @return void
 	 */
-	void processMessage(IEEE1588Port * port);
+	void processMessage( EtherPort *port );
+
 	/**
 	 * @brief  Gets the precise origin timestamp value
 	 * @return preciseOriginTimestamp value
@@ -887,7 +900,8 @@ public:
 	}
 
 	friend PTPMessageCommon *buildPTPMessage
-	(char *buf, int size, LinkLayerAddress * remote, IEEE1588Port * port);
+	( char *buf, int size, LinkLayerAddress *remote,
+	  EtherPort *port );
 };
 
 /**
@@ -910,22 +924,25 @@ class PTPMessagePathDelayReq : public PTPMessageCommon {
 	/**
 	 * @brief Builds the PTPMessagePathDelayReq message
 	 */
-	PTPMessagePathDelayReq(IEEE1588Port * port);
+	PTPMessagePathDelayReq( EtherPort *port );
 
 	/**
-	 * @brief  Assembles PTPMessagePathDelayReq message on the IEEE1588Port payload
-	 * @param  port IEEE1588Port where the message will be assembled
+	 * @brief  Assembles PTPMessagePathDelayReq message on the
+	 * EtherPort payload
+	 * @param  port EtherPort where the message will be
+	 * assembled
 	 * @param  destIdentity [in] Destination PortIdentity
 	 * @return void
 	 */
-	void sendPort(IEEE1588Port * port, PortIdentity * destIdentity);
+	void sendPort
+	( EtherPort *port, PortIdentity *destIdentity );
 
 	/**
 	 * @brief  Processes PTP messages
-	 * @param  port [in] IEEE1588Port
+	 * @param  port [in] EtherPort
 	 * @return void
 	 */
-	void processMessage(IEEE1588Port * port);
+	void processMessage( EtherPort *port );
 
 	/**
 	 * @brief  Gets origin timestamp value
@@ -936,7 +953,8 @@ class PTPMessagePathDelayReq : public PTPMessageCommon {
 	}
 
 	friend PTPMessageCommon *buildPTPMessage
-	(char *buf, int size, LinkLayerAddress * remote, IEEE1588Port * port);
+	( char *buf, int size, LinkLayerAddress *remote,
+	  EtherPort *port );
 };
 
 /**
@@ -957,22 +975,25 @@ public:
 	/**
 	 * @brief Builds the PTPMessagePathDelayResp object
 	 */
-	PTPMessagePathDelayResp(IEEE1588Port * port);
+	PTPMessagePathDelayResp( EtherPort *port );
 
 	/**
-	 * @brief  Assembles PTPMessagePathDelayResp message on the IEEE1588Port payload
-	 * @param  port IEEE1588Port where the message will be assembled
+	 * @brief  Assembles PTPMessagePathDelayResp message on the
+	 * EtherPort payload
+	 * @param  port EtherPort where the message will be
+	 * assembled
 	 * @param  destIdentity [in] Destination PortIdentity
 	 * @return void
 	 */
-	void sendPort(IEEE1588Port * port, PortIdentity * destIdentity);
+	void sendPort
+	( EtherPort *port, PortIdentity *destIdentity );
 
 	/**
 	 * @brief  Processes PTP messages
-	 * @param  port [in] IEEE1588Port
+	 * @param  port [in] EtherPort
 	 * @return void
 	 */
-	void processMessage(IEEE1588Port * port);
+	void processMessage( EtherPort *port );
 
 	/**
 	 * @brief  Sets the request receipt timestamp
@@ -1005,7 +1026,8 @@ public:
 	}
 
 	friend PTPMessageCommon *buildPTPMessage
-	(char *buf, int size, LinkLayerAddress * remote, IEEE1588Port * port);
+	( char *buf, int size, LinkLayerAddress *remote,
+	  EtherPort *port );
 };
 
 /**
@@ -1022,7 +1044,7 @@ public:
 	/**
 	 * @brief Builds the PTPMessagePathDelayRespFollowUp object
 	 */
-	PTPMessagePathDelayRespFollowUp(IEEE1588Port * port);
+	PTPMessagePathDelayRespFollowUp( EtherPort *port );
 
 	/**
 	 * @brief Destroys the PTPMessagePathDelayRespFollowUp object
@@ -1030,19 +1052,22 @@ public:
 	~PTPMessagePathDelayRespFollowUp();
 
 	/**
-	 * @brief  Assembles PTPMessageRespFollowUp message on the IEEE1588Port payload
-	 * @param  port IEEE1588Port where the message will be assembled
+	 * @brief  Assembles PTPMessageRespFollowUp message on the
+	 * EtherPort payload
+	 * @param  port EtherPort where the message will be
+	 * assembled
 	 * @param  destIdentity [in] Destination PortIdentity
 	 * @return void
 	 */
-	void sendPort(IEEE1588Port * port, PortIdentity * destIdentity);
+	void sendPort
+	( EtherPort *port, PortIdentity *destIdentity );
 
 	/**
 	 * @brief  Processes PTP messages
-	 * @param  port [in] IEEE1588Port
+	 * @param  port [in] EtherPort
 	 * @return void
 	 */
-	void processMessage(IEEE1588Port * port);
+	void processMessage( EtherPort *port );
 
 	/**
 	 * @brief  Sets the response origin timestamp
@@ -1075,7 +1100,8 @@ public:
 	}
 
 	friend PTPMessageCommon *buildPTPMessage
-	(char *buf, int size, LinkLayerAddress * remote, IEEE1588Port * port);
+	( char *buf, int size, LinkLayerAddress *remote,
+	  EtherPort *port );
 };
 
 /*Exact fit. No padding*/
@@ -1198,7 +1224,7 @@ public:
 	/**
 	 * @brief Builds the PTPMessageSignalling object
 	 */
-	PTPMessageSignalling(IEEE1588Port * port);
+	PTPMessageSignalling( EtherPort *port );
 
 	/**
 	 * @brief Destroys the PTPMessageSignalling object
@@ -1216,22 +1242,25 @@ public:
 
 	/**
 	 * @brief  Assembles PTPMessageSignalling message on the
-	 *         IEEE1588Port payload
-	 * @param  port IEEE1588Port where the message will be assembled
+	 *         EtherPort payload
+	 * @param  port EtherPort where the message will be
+	 * assembled
 	 * @param  destIdentity [in] Destination PortIdentity
 	 * @return void
 	 */
-	void sendPort(IEEE1588Port * port, PortIdentity * destIdentity);
+	void sendPort
+	( EtherPort *port, PortIdentity *destIdentity );
 
 	/**
 	 * @brief  Processes PTP messages
-	 * @param  port [in] IEEE1588Port
+	 * @param  port [in] EtherPort
 	 * @return void
 	 */
-	void processMessage(IEEE1588Port * port);
+	void processMessage( EtherPort *port );
 
 	friend PTPMessageCommon *buildPTPMessage
-	  (char *buf, int size, LinkLayerAddress * remote, IEEE1588Port * port);
+	( char *buf, int size, LinkLayerAddress *remote,
+	  EtherPort * port);
 };
 
 #endif
