+++ daemons/gptp/common/avbts_osipc.hpp
@@ -36,7 +36,7 @@
 
 #include <stdint.h>
 #include <ptptypes.hpp>
-#include <avbts_port.hpp>
+#include <ether_port.hpp>
 
 /**@file*/
