+++ daemons/gptp/common/avbts_osthread.hpp
@@ -47,6 +47,7 @@ typedef enum { osthread_ok, osthread_error } OSThreadExitCode;
  * @brief Provides the OSThreadExitCode callback format
  */
 typedef OSThreadExitCode(*OSThreadFunction) (void *);
+typedef void *OSThreadFunctionArg;
 
 /**
  * @brief Provides a generic interface for threads
@@ -81,7 +82,7 @@ public:
 	 * @brief Creates a new thread
 	 * @return Pointer to OSThread object
 	 */
-	virtual OSThread * createThread() = 0;
+	virtual OSThread * createThread() const = 0;
 
 	/**
 	 * @brief Destroys the new thread
