+++ daemons/gptp/linux/src/linux_hal_generic.hpp
@@ -137,8 +137,8 @@ public:
 	 * @return TRUE if got the time successfully, FALSE otherwise
 	 */
 	virtual bool HWTimestamper_gettime
-	( Timestamp *system_time, Timestamp *device_time, uint32_t *local_clock,
-	  uint32_t *nominal_clock_rate );
+	( Timestamp *system_time, Timestamp *device_time,
+	  uint32_t *local_clock, uint32_t *nominal_clock_rate ) const;
 
 	/**
 	 * @brief  Gets the TX timestamp from hardware interface
